pyproject.toml
[project]
name = "agentic-data-lab"
version = "0.1.0"
description = "Agentic ETL Pipeline with AI-powered analysis and visualization"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "google-generativeai>=0.3.0",
    "duckdb>=0.9.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "pyarrow>=14.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["etl", "orchestrator"]